refactor: table configuration 수정 class 생성
