+++ src/main/java/capstone/design/control_automation/report/util/hwp/fragments/style/BorderFillStyler.java
@@ -1,0 +1,44 @@
+package capstone.design.control_automation.report.util.hwp.fragments.style;
+
+import kr.dogfoot.hwplib.object.docinfo.BorderFill;
+import kr.dogfoot.hwplib.object.docinfo.DocInfo;
+import kr.dogfoot.hwplib.object.docinfo.borderfill.BackSlashDiagonalShape;
+import kr.dogfoot.hwplib.object.docinfo.borderfill.BorderThickness;
+import kr.dogfoot.hwplib.object.docinfo.borderfill.BorderType;
+import kr.dogfoot.hwplib.object.docinfo.borderfill.SlashDiagonalShape;
+import kr.dogfoot.hwplib.object.docinfo.borderfill.fillinfo.PatternFill;
+import kr.dogfoot.hwplib.object.docinfo.borderfill.fillinfo.PatternType;
+import org.springframework.stereotype.Component;
+
+@Component
+public class BorderFillStyler {
+
+    public void addBorderFillInfo(DocInfo docInfo) {
+        BorderFill bf = docInfo.addNewBorderFill();
+        bf.getProperty().set3DEffect(false);
+        bf.getProperty().setShadowEffect(false);
+        bf.getProperty().setSlashDiagonalShape(SlashDiagonalShape.None);
+        bf.getProperty().setBackSlashDiagonalShape(BackSlashDiagonalShape.None);
+        bf.getLeftBorder().setType(BorderType.Solid);
+        bf.getLeftBorder().setThickness(BorderThickness.MM0_5);
+        bf.getLeftBorder().getColor().setValue(0x0);
+        bf.getRightBorder().setType(BorderType.Solid);
+        bf.getRightBorder().setThickness(BorderThickness.MM0_5);
+        bf.getRightBorder().getColor().setValue(0x0);
+        bf.getTopBorder().setType(BorderType.Solid);
+        bf.getTopBorder().setThickness(BorderThickness.MM0_5);
+        bf.getTopBorder().getColor().setValue(0x0);
+        bf.getBottomBorder().setType(BorderType.Solid);
+        bf.getBottomBorder().setThickness(BorderThickness.MM0_5);
+        bf.getDiagonalBorder().setType(BorderType.None);
+        bf.getDiagonalBorder().setThickness(BorderThickness.MM0_7);
+
+        bf.getFillInfo().getType().setPatternFill(true);
+        bf.getFillInfo().createPatternFill();
+        PatternFill pf = bf.getFillInfo().getPatternFill();
+        pf.setPatternType(PatternType.None);
+        pf.getBackColor().setValue(-1);
+        pf.getPatternColor().setValue(0);
+    }
+
+}
